Renamed ml module and command to mlalpha (#72)
Renamed ml module and command to mlalpha
