+++ datalab/mlalpha/commands/__init__.py
@@ -13,5 +13,5 @@
 
 from __future__ import absolute_import
 
-from . import _ml
+from . import _mlalpha
 from . import _tensorboard
