+++ setup.py
@@ -30,8 +30,8 @@ setup(
     'datalab.data',
     'datalab.data.commands',
     'datalab.kernel',
-    'datalab.ml',
-    'datalab.ml.commands',
+    'datalab.mlalpha',
+    'datalab.mlalpha.commands',
     'datalab.notebook',
     'datalab.stackdriver',
     'datalab.stackdriver.commands',
